Plato BOT
This program can be used to clock in hours required to pass a course. NOTE: Courses offer exams and other grading materials. This does not interfere with that in any way. This program simply logs in/out, moves the mouse, and selects the next page, to prevent inactivity timeouts. This tool was made for educational purposes ONLY.